Revert "Factor out Test packages to testsuites"
This breaks feldspar-compiler in a non-fixable
way so revert for now.

This reverts commit aee836cca37f9ab8ee9fb2f868753f88237c384c.
