+++ feldspar-language.cabal
@@ -89,11 +89,14 @@ library
     bytestring                  >= 0.10   && < 0.11,
     containers                  >= 0.4    && < 0.7,
     mtl                         >= 2.0    && < 2.3,
+    QuickCheck                  >= 2.7    && < 3,
     patch-combinators           >= 0.2    && < 0.3,
     prelude-edsl                             < 0.4,
     template-haskell,
     tuple                       >= 0.2    && < 0.5,
     monad-par                   >= 0.3.4.5,
+    deepseq,
+    random                      >= 1.0    && < 1.2,
     data-default                >= 0.5.3  && < 0.8,
     data-hash                   >= 0.2.0.1 && < 100,
     tree-view
@@ -110,7 +113,6 @@ test-suite range
   main-is: RangeTest.hs
 
   other-modules:
-    Feldspar.Core.TestUtils
     Feldspar.Range.Test
 
   default-language: Haskell2010
@@ -118,10 +120,8 @@ test-suite range
   build-depends:
     feldspar-language,
     base,
-    data-hash        >= 0.2.0.1 && < 100,
-    deepseq,
-    QuickCheck       >= 2.7 && < 3,
     random           >= 1   && < 2,
+    QuickCheck       >= 2.7 && < 3,
     tasty            >= 0.3,
     tasty-quickcheck >= 0.2
 
@@ -136,7 +136,6 @@ test-suite semantics
 
   other-modules:
     Feldspar.Core.Test
-    Feldspar.Core.TestUtils
     Feldspar.Stream.Test
     Feldspar.Vector.Test
     Feldspar.Mutable.Test
@@ -147,10 +146,7 @@ test-suite semantics
   build-depends:
     feldspar-language,
     base,
-    data-hash        >= 0.2.0.1 && < 100,
-    deepseq,
     QuickCheck       >= 2.7 && < 3,
-    random           >= 1   && < 2,
     tasty            >= 0.3,
     tasty-th         >= 0.1,
     tasty-quickcheck >= 0.2
