+++ src/Feldspar/Range.hs
@@ -46,7 +46,10 @@ import Data.Int
 import Data.Word
 import Data.Default
 import Data.Hash
+import qualified Test.QuickCheck as Q
+import System.Random (Random(..))
 
+import Control.DeepSeq (NFData(..))
 import Foreign.Storable (Storable)
 import Language.Haskell.TH.Syntax (Lift(..))
 
@@ -80,14 +83,16 @@ instance (Show a, Bounded a, Eq a) => Show (Range a)
 newtype WordN = WordN Word32
   deriving
     ( Eq, Ord, Num, Enum, IO.Ix, Real, Integral, Bits, Bounded
-    , Storable, Default, FiniteBits, Hashable, Lift
+    , Q.Arbitrary, Random, Storable, NFData, Default
+    , FiniteBits, Hashable, Lift
     )
 
 -- | Target-dependent signed integers
 newtype IntN = IntN Int32
   deriving
     ( Eq, Ord, Num, Enum, IO.Ix, Real, Integral, Bits, Bounded
-    , Storable, Default, FiniteBits, Hashable
+    , Q.Arbitrary, Random, Storable, NFData, Default
+    , FiniteBits, Hashable
     )
 
 instance Show WordN
