+++ tests/Feldspar/Core/Test.hs
@@ -12,7 +12,6 @@ import qualified Data.List as DL
 
 import Feldspar
 import Feldspar.Mutable
-import Feldspar.Core.TestUtils
 
 tests = $(testGroupGenerator)
